Trying to track down the color palette change on Configuration save. Replacing unnecessary (byte) conversions. Fixed lost keyboard layout.

+++ VCCSharp/Modules/Events.cs
@@ -55,7 +55,7 @@ namespace VCCSharp.Modules
             {
                 _modules.Clipboard.Abort = true; //--Abort Pasting if happening
 
-                _modules.Emu.ResetPending = (byte)state;
+                _modules.Emu.ResetPending = state;
             }
         }
